Hofatrano-backend/requirements.txt
Django==5.1.7
djangorestframework==3.15.2
django-cors-headers==4.6.0
reportlab==4.2.5
psycopg2-binary==2.9.9
python-dotenv==1.2.2
